Fixup ActiveRecord main workflow
Rails switched to `sqlite3 >= 2.0` in [0], so workflow fails due to
the current `~> 1.3` constraint. It's quite complicated to make
current gemfiles setup to handle this, because `sqlite3` is not an AR
dependency, but loaded when needed, so we need to edit all the
gemfiles with explicit sqlite3 constraint.
Introduce Appraisal to generate gemfiles, so it's easier to handle.

[0] https://github.com/rails/rails/commit/2976d3767e6572ee1838b50b5f401983918f99c7